Bug 81313 - Zimlet JS Error - parentNode is not defined
Add extra check for parentNode existence.

(From Lars P. Thomsen.)

http://bugzilla.zimbra.com/show_bug.cgi?id=81313

+++ ZimbraWebClient/WebRoot/js/ajax/dwt/widgets/DwtMenu.js
@@ -915,7 +915,7 @@ function(child) {
 			this._table.rows[0].deleteCell(Dwt.getCellIndex(cell));
 		} else {
 			var el = child.getHtmlElement();
-			if (el && el.parentNode.parentNode.rowIndex > -1)// Make sure that the element exists in the table
+			if (el && el.parentNode && el.parentNode.parentNode.rowIndex > -1)// Make sure that the element exists in the table
 				this._table.deleteRow(el.parentNode.parentNode.rowIndex);
 		}
 	}
